scripts/skin-recolor/requirements.txt
numpy==2.0.0
UnityPy==1.10.6
pillow==10.4.0
